% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datapacakge_init}
\alias{datapacakge_init}
\title{Initialise a data package from a data frame, metadata list, and source code
file used to create the data set.}
\usage{
datapacakge_init(df, meta, source_clean)
}
\arguments{
\item{df}{The data frame object name of the data frame you would like to convert
into a data package.}

\item{meta}{The list object name with the data frames meta data. The list
item names must conform to the Open Knowledge Foundation's Data Package
Protocol (see \url{http://dataprotocols.org/data-packages/}).}

\item{source_clean}{a character string file path pointing to the source code
file used to gather and clean the \code{df} data frame. Can be in R or any
other language, e.g. Python. \code{source_clean} is not required, but HIGHLY
RECOMMENDED.}
}
\description{
Initialise a data package from a data frame, metadata list, and source code
file used to create the data set.
}

