% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datapackage_info.R
\name{datapackage_info}
\alias{datapackage_info}
\title{Return key meta information about the data package}
\usage{
datapackage_info(path, as_list = FALSE)
}
\arguments{
\item{path}{character string file path to the data package. If empty, then
the datapackage.json meta data file is searched for in the working directory.
Can also accept a datapackage.json file parsed in R as a list.}

\item{as_list}{logical indicating whether or not to return the
datapackage.json file as a list.}
}
\description{
Return key meta information about the data package
}
\examples{
\dontrun{
# Print information when working directory is a data package
datapackage_info()
}
}

