% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datapackage_init}
\alias{datapackage_init}
\title{Initialise a data package from a data frame, metadata list, and source code
file used to create the data set.}
\usage{
datapackage_init(df, package_name, meta = NULL, source_cleaner = NULL)
}
\arguments{
\item{df}{The data frame object name of the data frame you would like to convert
into a data package.}

\item{package_name}{character string name for the data package. Unnecessary
if the \code{name} field is specified in \code{meta}.}

\item{meta}{The list object name with the data frames meta data. The list
item names must conform to the Open Knowledge Foundation's Data Package
Protocol (see \url{http://dataprotocols.org/data-packages/}). If
\code{meta = NULL} then a barebones \code{datapackage.json} file will be
created.}

\item{source_cleaner}{a character string or vector of file paths pointing to
the source code file used to gather and clean the \code{df} data frame. Can
be in R or any other language, e.g. Python. Following Data Package convention
the scripts are renamed \code{process*.*}. You can also  \code{source_cleaner} is not
required, but HIGHLY RECOMMENDED.}
}
\description{
Initialise a data package from a data frame, metadata list, and source code
file used to create the data set.
}

