% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datapackage_install}
\alias{datapackage_install}
\title{Install a data package}
\usage{
datapackage_install(path, load_file = NULL)
}
\arguments{
\item{path}{character string path to the data package directory. Can be a
local directory, a remote repository's URL, or a GitHub username/repo.}

\item{load_file}{character string specifying the name of the data file to
load into R.
Note: unfortunately R only allows one file to be returned at a time.}
}
\description{
Install a data package
}

