% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meta_template}
\alias{meta_template}
\title{Template for datapackage.json}
\usage{
meta_template(df, name, data_paths)
}
\arguments{
\item{df}{The data frame object name of the data frame you would like to
convert into a data package.}

\item{name}{character string name of the datapackage.}

\item{data_paths}{character vector of df paths.}
}
\description{
Template for datapackage.json
}
\keyword{helpers}

